pyproject.toml
[project]
name = "webcontentgenerationpipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.86.0",
    "pydantic>=2.11.6",
    "python-dotenv>=1.1.0",
]
